suggested changes and improvements
