+++ .gitignore
@@ -14,6 +14,7 @@ node/
 .setting
 .project
 .classpath
+.gradle
 release.properties
 *.releaseBackup
