+++ forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/payment/v3_1_5/domesticpayments/DomesticPaymentConsentsApiController.java
@@ -58,8 +58,8 @@ import javax.servlet.http.HttpServletRequest;
 import java.security.Principal;
 import java.util.Optional;
 
-import static com.forgerock.openbanking.common.model.openbanking.persistence.payment.converter.v3_1_5.ResponseConverter.ReadRefundAccountConverter.toOBWriteDomesticConsentResponse5DataReadRefundAccount;
-import static com.forgerock.openbanking.common.model.openbanking.persistence.payment.converter.v3_1_5.ResponseConverter.StatusCodeConverter.toOBWriteDomesticConsentResponse5DataStatus;
+import static com.forgerock.openbanking.common.model.openbanking.persistence.payment.converter.v3_1_5.ResponseReadRefundAccountConverter.toOBWriteDomesticConsentResponse5DataReadRefundAccount;
+import static com.forgerock.openbanking.common.model.openbanking.persistence.payment.converter.v3_1_5.ResponseStatusCodeConverter.toOBWriteDomesticConsentResponse5DataStatus;
 import static com.forgerock.openbanking.common.services.openbanking.IdempotencyService.validateIdempotencyRequest;
 import static com.forgerock.openbanking.common.services.openbanking.converter.common.FRAccountIdentifierConverter.toOBDebtorIdentification1;
 import static com.forgerock.openbanking.common.services.openbanking.converter.payment.FRDataAuthorisationConverter.toOBWriteDomesticConsent4DataAuthorisation;
@@ -115,7 +115,7 @@ public class DomesticPaymentConsentsApiController implements DomesticPaymentCons
         if (consentByIdempotencyKey.isPresent()) {
             validateIdempotencyRequest(xIdempotencyKey, frWriteDomesticConsent, consentByIdempotencyKey.get(), () -> consentByIdempotencyKey.get().getDomesticConsent());
             log.info("Idempotent request is valid. Returning [201 CREATED] but take no further action.");
-            return ResponseEntity.status(HttpStatus.CREATED).body(entityInstance(consentByIdempotencyKey.get()));
+            return ResponseEntity.status(HttpStatus.CREATED).body(responseEntity(consentByIdempotencyKey.get()));
         }
         log.debug("No consent with matching idempotency key has been found. Creating new consent.");
 
@@ -133,7 +133,7 @@ public class DomesticPaymentConsentsApiController implements DomesticPaymentCons
         consentMetricService.sendConsentActivity(new ConsentStatusEntry(domesticConsent.getId(), domesticConsent.getStatus().name()));
         domesticConsent = domesticConsentRepository.save(domesticConsent);
         log.info("Created consent id: '{}'", domesticConsent.getId());
-        return ResponseEntity.status(HttpStatus.CREATED).body(entityInstance(domesticConsent));
+        return ResponseEntity.status(HttpStatus.CREATED).body(responseEntity(domesticConsent));
     }
 
     @Override
@@ -153,7 +153,7 @@ public class DomesticPaymentConsentsApiController implements DomesticPaymentCons
         }
         FRDomesticConsent domesticConsent = isDomesticConsent.get();
 
-        return ResponseEntity.ok(entityInstance(domesticConsent));
+        return ResponseEntity.ok(responseEntity(domesticConsent));
     }
 
     @Override
@@ -193,7 +193,7 @@ public class DomesticPaymentConsentsApiController implements DomesticPaymentCons
 
     }
 
-    private OBWriteDomesticConsentResponse5 entityInstance(FRDomesticConsent domesticConsent) {
+    private OBWriteDomesticConsentResponse5 responseEntity(FRDomesticConsent domesticConsent) {
         return new OBWriteDomesticConsentResponse5()
                 .data(new OBWriteDomesticConsentResponse5Data()
                         .readRefundAccount(toOBWriteDomesticConsentResponse5DataReadRefundAccount(domesticConsent.getDomesticConsent().getData().getReadRefundAccount()))
