+++ forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/payment/v3_1_5/domesticstandingorders/DomesticStandingOrderConsentsApiController.java
@@ -54,8 +54,8 @@ import javax.servlet.http.HttpServletRequest;
 import java.security.Principal;
 import java.util.Optional;
 
-import static com.forgerock.openbanking.common.model.openbanking.persistence.payment.converter.v3_1_5.ResponseConverter.ReadRefundAccountConverter.toOBWriteDomesticStandingOrderConsentResponse6DataReadRefundAccount;
-import static com.forgerock.openbanking.common.model.openbanking.persistence.payment.converter.v3_1_5.ResponseConverter.StatusCodeConverter.toOBWriteDomesticStandingOrderConsentResponse6DataStatus;
+import static com.forgerock.openbanking.common.model.openbanking.persistence.payment.converter.v3_1_5.ResponseReadRefundAccountConverter.toOBWriteDomesticStandingOrderConsentResponse6DataReadRefundAccount;
+import static com.forgerock.openbanking.common.model.openbanking.persistence.payment.converter.v3_1_5.ResponseStatusCodeConverter.toOBWriteDomesticStandingOrderConsentResponse6DataStatus;
 import static com.forgerock.openbanking.common.services.openbanking.IdempotencyService.validateIdempotencyRequest;
 import static com.forgerock.openbanking.common.services.openbanking.converter.common.FRAccountIdentifierConverter.toOBDebtorIdentification1;
 import static com.forgerock.openbanking.common.services.openbanking.converter.payment.FRDataAuthorisationConverter.toOBWriteDomesticConsent4DataAuthorisation;
@@ -104,7 +104,7 @@ public class DomesticStandingOrderConsentsApiController implements DomesticStand
         if (consentByIdempotencyKey.isPresent()) {
             validateIdempotencyRequest(xIdempotencyKey, frStandingOrderConsent, consentByIdempotencyKey.get(), () -> consentByIdempotencyKey.get().getDomesticStandingOrderConsent());
             log.info("Idempotent request is valid. Returning [201 CREATED] but take no further action.");
-            return ResponseEntity.status(HttpStatus.CREATED).body(entityInstance(consentByIdempotencyKey.get()));
+            return ResponseEntity.status(HttpStatus.CREATED).body(responseEntity(consentByIdempotencyKey.get()));
         }
         log.debug("No consent with matching idempotency key has been found. Creating new consent...");
 
@@ -121,7 +121,7 @@ public class DomesticStandingOrderConsentsApiController implements DomesticStand
         consentMetricService.sendConsentActivity(new ConsentStatusEntry(domesticStandingOrderConsent.getId(), domesticStandingOrderConsent.getStatus().name()));
         domesticStandingOrderConsent = domesticStandingOrderConsentRepository.save(domesticStandingOrderConsent);
         log.info("Created consent id: '{}'", domesticStandingOrderConsent.getId());
-        return ResponseEntity.status(HttpStatus.CREATED).body(entityInstance(domesticStandingOrderConsent));
+        return ResponseEntity.status(HttpStatus.CREATED).body(responseEntity(domesticStandingOrderConsent));
     }
 
     public ResponseEntity getDomesticStandingOrderConsentsConsentId(
@@ -140,10 +140,10 @@ public class DomesticStandingOrderConsentsApiController implements DomesticStand
         }
         FRDomesticStandingOrderConsent domesticStandingOrderConsent = isDomesticStandingOrderConsent.get();
 
-        return ResponseEntity.ok(entityInstance(domesticStandingOrderConsent));
+        return ResponseEntity.ok(responseEntity(domesticStandingOrderConsent));
     }
 
-    private OBWriteDomesticStandingOrderConsentResponse6 entityInstance(FRDomesticStandingOrderConsent domesticStandingOrderConsent) {
+    private OBWriteDomesticStandingOrderConsentResponse6 responseEntity(FRDomesticStandingOrderConsent domesticStandingOrderConsent) {
         return new OBWriteDomesticStandingOrderConsentResponse6()
                 .data(new OBWriteDomesticStandingOrderConsentResponse6Data()
                         .readRefundAccount(toOBWriteDomesticStandingOrderConsentResponse6DataReadRefundAccount(domesticStandingOrderConsent.getDomesticStandingOrderConsent().getData().getReadRefundAccount()))
