+++ forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/test/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/payment/v3_1_6/internationalstandingorders/InternationalStandingOrdersApiControllerIT.java
@@ -61,8 +61,8 @@ import java.util.UUID;
 import static com.forgerock.openbanking.aspsp.rs.store.api.openbanking.testsupport.domain.FRAmountTestDataFactory.aValidFRAmount;
 import static com.forgerock.openbanking.aspsp.rs.store.api.openbanking.testsupport.domain.FRRiskTestDataFactory.aValidFRRisk;
 import static com.forgerock.openbanking.common.services.openbanking.converter.payment.FRPaymentRiskConverter.toOBRisk1;
-import static com.forgerock.openbanking.common.services.openbanking.converter.payment.FRWriteInternationalStandingOrderConsentConverter.toOBWriteInternationalStandingOrder4;
 import static com.forgerock.openbanking.common.services.openbanking.converter.payment.FRWriteInternationalStandingOrderConsentConverter.toOBWriteInternationalStandingOrder4DataInitiation;
+import static com.forgerock.openbanking.common.services.openbanking.converter.payment.FRWriteInternationalStandingOrderConverter.toFRWriteInternationalStandingOrder;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
@@ -162,7 +162,7 @@ public class InternationalStandingOrdersApiControllerIT {
         FRInternationalStandingOrderPaymentSubmission submission = submissionRepository.findById(response.getBody().getData().getInternationalStandingOrderId()).get();
         assertThat(submission.getId()).isEqualTo(consentResponse.getData().getConsentId());
         assertThat(consentResponse.getData().getRefund()).isNull();
-        assertThat(toOBWriteInternationalStandingOrder4(submission.getInternationalStandingOrder())).isEqualTo(submissionRequest);
+        assertThat(submission.getInternationalStandingOrder()).isEqualTo(toFRWriteInternationalStandingOrder(submissionRequest));
         assertThat(submission.getVersion()).isEqualTo(OBVersion.v3_1_6);
     }
 
@@ -194,7 +194,7 @@ public class InternationalStandingOrdersApiControllerIT {
         FRInternationalStandingOrderPaymentSubmission submission = submissionRepository.findById(response.getBody().getData().getInternationalStandingOrderId()).get();
         assertThat(submission.getId()).isEqualTo(consentResponse.getData().getConsentId());
         assertThat(consentResponse.getData().getRefund()).isNull();
-        assertThat(toOBWriteInternationalStandingOrder4(submission.getInternationalStandingOrder())).isEqualTo(submissionRequest);
+        assertThat(submission.getInternationalStandingOrder()).isEqualTo(toFRWriteInternationalStandingOrder(submissionRequest));
         assertThat(submission.getVersion()).isEqualTo(OBVersion.v3_1_6);
     }
 
@@ -234,8 +234,7 @@ public class InternationalStandingOrdersApiControllerIT {
         assertThat(refund.getAccount().getName()).isEqualTo(consent.getInitiation().getDebtorAccount().getName());
         assertThat(refund.getAccount().getSchemeName()).isEqualTo(consent.getInitiation().getDebtorAccount().getSchemeName());
         assertThat(refund.getAccount().getSecondaryIdentification()).isEqualTo(consent.getInitiation().getDebtorAccount().getSecondaryIdentification());
-
-        assertThat(toOBWriteInternationalStandingOrder4(submission.getInternationalStandingOrder())).isEqualTo(submissionRequest);
+        assertThat(submission.getInternationalStandingOrder()).isEqualTo(toFRWriteInternationalStandingOrder(submissionRequest));
         assertThat(submission.getVersion()).isEqualTo(OBVersion.v3_1_6);
     }
